ai_eng/LangChain_demo_1/pyproject.toml
[project]
name = "langchain-demo-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
